포지션 쪽 컬럼명, 형식 수정

+++ backend/jobplatform/src/main/java/jobplatform/fo/enterprise/domain/dto/ApplyConditionDataDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ApplyConditionDataDTO {
 	//지원 순번
-	private int apy_sq;
+	private Long apy_sq;
 	//지원 상태
 	private String apy_cndtn;
 	//면접 일시
